backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
python-dotenv
firebase-admin
